% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/chronic_carriers.R
\name{chronic_carriers_lhs}
\alias{chronic_carriers_lhs}
\title{Draw latin hypercube samples for the trypanosomiasis model with chronic carriers}
\usage{
chronic_carriers_lhs(nsamples = 1, nruns = 100, seed = NULL)
}
\arguments{
\item{nsamples}{}

\item{nruns}{number of runs}

\item{seed}{random number seed}

\item{nsamples}{number of samples}
}
\value{
samples, posteriors, random numbers
}
\description{
Draw latin hypercube samples for the trypanosomiasis model with chronic carriers
}
\author{
Sebastian Funk
}

