% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/chronic_carriers.R
\name{chronic_carriers_prior}
\alias{chronic_carriers_prior}
\title{Draw prior  samples for the trypanosomiasis model with chronic carriers}
\usage{
chronic_carriers_prior(nsamples = 1, finite = FALSE, nruns = 100,
  seed = NULL, verbose = FALSE, passive = TRUE)
}
\arguments{
\item{nsamples}{number of samples}

\item{finite}{interpret \code{nsamples} as number of finite samples}

\item{nruns}{number of runs}

\item{seed}{random number seed}

\item{verbose}{whether to print out posteriors as it goes along}

\item{passive}{whether to fit to passive case data}
}
\value{
samples, posteriors, random numbers
}
\description{
Draw prior  samples for the trypanosomiasis model with chronic carriers
}
\author{
Sebastian Funk
}

