% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chronic_carriers.R
\name{chronic_carriers_sample}
\alias{chronic_carriers_sample}
\title{Draw samples for the trypanosomiasis model with chronic carriers}
\usage{
chronic_carriers_sample(nsamples = 1, seed, verbose = FALSE,
  passive = TRUE, calc.posterior = TRUE, villages, progress.bar = TRUE,
  sample = c("lhs", "prior"), ...)
}
\arguments{
\item{nsamples}{number of samples}

\item{seed}{random number seed}

\item{verbose}{whether to print out posteriors as it goes along}

\item{passive}{wheter to fit to passive detection data}

\item{calc.posterior}{whether to calculate the posterior. Alernatively, will calculate summary statistic}

\item{villages}{villages to simuate}

\item{progress.bar}{whether to show a progress bar (TRUE by default)}

\item{sample}{how to sample (lhs or prior)}

\item{...}{further options for rprior}
}
\value{
samples, posteriors, random numbers
}
\description{
Draw samples for the trypanosomiasis model with chronic carriers
}
\author{
Sebastian Funk
}

