% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/chronic_carriers.R
\name{dinit}
\alias{dinit}
\title{Evaluate the probability density of a given set of initial conditions of the trypanosomiasis model}
\usage{
dinit(init, village_number, theta, log = FALSE)
}
\arguments{
\item{init}{initial conditions}

\item{theta}{parameter vector}

\item{log}{whether to return the logarithm}

\item{village_data}{Village data}
}
\value{
Probability density
}
\description{
Evaluate the probability density of a given set of initial conditions of the trypanosomiasis model
}
\author{
Sebastian Funk
}

