% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/chronic_carriers.R
\name{dprior}
\alias{dprior}
\title{Evaluate prior probability density of a parameter vector for the trypanosomiasis model}
\usage{
dprior(theta, log = FALSE)
}
\arguments{
\item{theta}{Parameter vector}

\item{log}{whether to take return the logarithm of the prior probability density}
}
\value{
Probability density
}
\description{
Evaluate prior probability density of a parameter vector for the trypanosomiasis model
}
\author{
Sebastian Funk
}

