% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chronic_carriers.R
\name{likelihood}
\alias{likelihood}
\title{Evaluate the likelihood of a model state of the trypanosomiasis model}
\usage{
likelihood(state, stage1_detected, stage2_detected, theta, attendance = 1,
  active = TRUE, log = FALSE)
}
\arguments{
\item{state}{Model state}

\item{stage1_detected}{Number of stage 1 cases detected}

\item{stage2_detected}{Number of stage 2 cases detected}

\item{theta}{parameter vector}

\item{attendance}{screening attendance}

\item{log}{whether to return the logarithm}
}
\value{
point likelihood
}
\description{
Evaluate the likelihood of a model state of the trypanosomiasis model
}
\author{
Sebastian Funk
}

