% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/chronic_carriers.R
\name{rprior}
\alias{rprior}
\title{Draw a parameter sample from the prior density}
\usage{
rprior(villages = 1)
}
\arguments{
\item{villages}{Number of villages with different parameters}
}
\value{
parameter vector
}
\description{
Draw a parameter sample from the prior density
}
\author{
Sebastian Funk
}

