% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/chronic_carriers.R
\name{sim_chronic_carriers}
\alias{sim_chronic_carriers}
\title{Simulate the trypanosomiasis model, with passive case detection}
\usage{
sim_chronic_carriers(init, theta, village_number)
}
\arguments{
\item{init}{initial state}

\item{theta}{parameter vector}

\item{village_number}{village number}
}
\value{
trajectory
}
\description{
Simulate the trypanosomiasis model, with passive case detection
}
\author{
Sebastian Funk
}

