% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/chronic_carriers.R
\name{sim_chronic_rates}
\alias{sim_chronic_rates}
\title{Simulation step for trypanosomiasis model with chronic infections}
\usage{
sim_chronic_rates(state, theta, t)
}
\arguments{
\item{state}{current state of the model}

\item{theta}{parameter vector of the model}

\item{t}{time}
}
\value{
current transition rates
}
\description{
For use with \code{ssa.adaptivetau}
}
\author{
Sebastian Funk
}

