% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chronic_carriers.R
\name{sim_trajectory}
\alias{sim_trajectory}
\title{Produe a single simuation run}
\usage{
sim_trajectory(theta, init, village)
}
\arguments{
\item{theta}{parameter vector}
}
\value{
data frame with trajctor
}
\description{
Produe a single simuation run
}
\author{
seb
}

