% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/chronic_carriers.R
\name{traj.likelihood}
\alias{traj.likelihood}
\title{Evaluate the likelihood of a trajectory or the trypanosomiasis model}
\usage{
traj.likelihood(traj, theta, village_number, log = FALSE)
}
\arguments{
\item{traj}{Model trajectory}

\item{theta}{parameter vector}

\item{log}{whether to return the logarithm}

\item{attendance}{screening attendance}
}
\value{
trajectory likelihood
}
\description{
Evaluate the likelihood of a trajectory or the trypanosomiasis model
}
\author{
Sebastian Funk
}

