% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chronic_carriers.R
\name{traj_likelihood}
\alias{traj_likelihood}
\title{Calculate the posterior density for a given set of parameters for the trypanosomiasis model}
\usage{
traj_likelihood(theta, village_number, nruns, log = FALSE, ...)
}
\arguments{
\item{theta}{parameter vector}

\item{nruns}{number of runs (to estimate the likelihood)}

\item{log}{whether to return the logarithm of the posterior density}

\item{...}{}

\item{village_data}{village data}

\item{cum_data}{cumulative data}
}
\value{
posterior density
}
\description{
Calculate the posterior density for a given set of parameters for the trypanosomiasis model
}
\author{
Sebastian Funk
}

