% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chronic_carriers.R
\name{chronic_carriers_mcmc}
\alias{chronic_carriers_mcmc}
\title{run MCMC on the model for chronic carriers of trypanosomiasis}
\usage{
chronic_carriers_mcmc(init, n_iterations, sd, upper, lower, epsilon,
  data_summary, verbose = FALSE, ...)
}
\arguments{
\item{init}{initial state of the chain}

\item{n_iterations}{number of iterations}

\item{sd}{standard deviation of each parameter}

\item{upper}{upper limit on parameters (if given)}

\item{lower}{lower limit on parameters (if given)}

\item{epsilon}{value of epsilon (if ABC is run)}

\item{data_summary}{summary statistic of the data (if ABC is run)}

\item{verbose}{whether to print verbose output}

\item{...}{parameters to be passed to \code{param_posterior_villages}}
}
\value{
chain
}
\description{
run MCMC on the model for chronic carriers of trypanosomiasis
}
\author{
Sebastian Funk
}

