% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/chronic_carriers.R
\name{param_posterior_villages}
\alias{param_posterior_villages}
\title{Calculate the posterior density for a given set of parameters for the trypanosomiasis model, using a model that encompasses all villages}
\usage{
param_posterior_villages(theta, nruns, log = FALSE, ...)
}
\arguments{
\item{theta}{parameter vector}

\item{nruns}{number of runs (to estimate the likelihood)}

\item{log}{whether to return the logarithm of the posterior density}

\item{...}{}

\item{village_screening}{village screening data}

\item{cum_data}{village cumulative data}
}
\value{
posterior density
}
\description{
Calculate the posterior density for a given set of parameters for the trypanosomiasis model, using a model that encompasses all villages
}
\author{
Sebastian Funk
}

