% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chronic_carriers.R
\name{param_posterior_villages}
\alias{param_posterior_villages}
\title{Calculate the posterior density for a given set of parameters for the trypanosomiasis model, using a model that encompasses all villages}
\usage{
param_posterior_villages(theta, villages, log = TRUE, ...)
}
\arguments{
\item{theta}{parameter vector}

\item{villages}{villages to simuate}

\item{log}{whether to return the logarithm}

\item{...}{further options for traj_likelihood}
}
\value{
posterior density
}
\description{
Calculate the posterior density for a given set of parameters for the trypanosomiasis model, using a model that encompasses all villages
}
\author{
Sebastian Funk
}

