% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chronic_carriers.R
\name{param_sumstat_villages}
\alias{param_sumstat_villages}
\title{Calculate summarystatistic for a given set of parameters for the trypanosomiasis model, using a model that encompasses all villages}
\usage{
param_sumstat_villages(theta, nruns, villages, ...)
}
\arguments{
\item{theta}{parameter vector}

\item{nruns}{number of runs}

\item{villages}{villages to simuate}

\item{log}{whether to return the logarithm of the posterior density}
}
\value{
posterior density
}
\description{
Calculate summarystatistic for a given set of parameters for the trypanosomiasis model, using a model that encompasses all villages
}
\author{
Sebastian Funk
}

