% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/chronic_carriers.R
\name{rinit}
\alias{rinit}
\title{Draw an initial condition sample from the prior density distribution of the trypanosomiasis model}
\usage{
rinit(theta, rand = NULL, equilibrium = TRUE, village_number = 1)
}
\arguments{
\item{theta}{Parameter vector}

\item{rand}{A list of combinations for how infections are distributed between chronic and symptomatic infections, with corresponding probabilities}

\item{equilibrium}{whether the initial conditions should be derived from equilibrium states}
}
\value{
initial conditions vector
}
\description{
Draw an initial condition sample from the prior density distribution of the trypanosomiasis model
}
\author{
Sebastian Funk
}

