% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chronic_carriers.R
\name{rprior}
\alias{rprior}
\title{Draw a parameter sample from the prior density}
\usage{
rprior(villages = 1, passive = TRUE, background = TRUE,
  transmitted = FALSE, chronic = TRUE)
}
\arguments{
\item{villages}{Number of villages with different parameters}

\item{passive}{whether to simulate passive detection}

\item{background}{whether to simuate backgroud infection rates}

\item{transmitted}{whether to simulate transmitted infection rates}

\item{chronic}{whether chronic carriers transmit}
}
\value{
parameter vector
}
\description{
Default arameters reflect Rebecca's study
}
\author{
Sebastian Funk
}

